Convert `DialogConfirmation` to TS
